+++ R/zzz.R
@@ -14,6 +14,11 @@
 #' @section Package options:
 #'
 #' \itemize{
+#'
+#'   \item \code{options('huxtable.add_colnames')} sets the default value for \code{add_colnames} in
+#'     \code{\link{huxtable}} and \code{\link{as_huxtable}}. If it is unset, \code{add_colnames} defaults to
+#'     \code{FALSE}; in a future release, the default will become \code{TRUE}.
+#'
 #'   \item \code{options('huxtable.print')} sets the print method for huxtable objects. See \code{\link{print.huxtable}}.
 #'   \item \code{options('huxtable.color_screen')}. If \code{TRUE} and package \code{crayon} is available, huxtables
 #'   will be printed in color on screen.
@@ -23,6 +28,7 @@
 #'
 #'   \item \code{options('huxtable.knit_print_df_theme')}. A one-argument function applied to theme the huxtableized
 #'   data frame before printing in knitr. Defaults to \code{\link{theme_plain}}.
+#'
 #' }
 #'
 #'
@@ -38,4 +44,10 @@ NULL
     huxtable.knit_print_df_theme = getOption('huxtable.knit_print_df_theme', theme_plain),
     huxtable.color_screen        = getOption('huxtable.color_screen', requireNamespace('crayon', quietly = TRUE))
   )
+
+  if (is.null(getOption("huxtable.add_colnames"))) packageStartupMessage(
+          'By default, add_colnames = FALSE in huxtable and as_huxtable.default.\n',
+          'This will change in a future release. To suppress this message, \n',
+          'set `options("huxtable.add_colnames")` to TRUE or FALSE.'
+        )
 }